Syukri branch

+++ trace.c
@@ -7,11 +7,13 @@ int trace(Matrix a)
 		exit(EXIT_FAILURE);	
 	}
 
-	float sum = 0;
+	double sum = 0;
 
 	for(int i = 0;i < a.value_num; i++) {
 		if(a.data[i].row == a.data[i].col){
 			sum = sum + a.data[i].value;
+			//printf("The value is: %f\n", a.data[i].value);
+			//printf("Sum is at: %f\n", sum);
 		}
 	
 	}
